requirements.txt

Flask==0.12
itsdangerous==0.24
Jinja2==2.8
requests==2.8.1
Werkzeug==0.11
Flask-SQLAlchemy==2.1
python-dateutil==2.2
psycopg2==2.6.1
nose==1.3.7
PyJWT==1.4.1
cryptography==1.6
SQLAlchemy==1.0.9
SQLAlchemy-Utils==0.31.1
coverage==4.0.3
blinker==1.4
